Refactor initial block creation and block down motion
Create the block offscreen so that it doesn't immediately appear.
Change down motion to using an alarm. It will move a whole grid movement
on alarm event trigger, the speed will be manageable through the alarm
timing.
Change the down key to immediately trigger this alarm instead of directly
modifying the block's y position.