Refactor block rotation
Instead of having separate subimages for every rotation, actually
rotate the blocks. The reason that I had done it that way previously
was to fix a perceived issue with how rotation worked in this game.

However, doing things this way caused issues with collision masks
and was causing issues with the collision detection. Furthermore,
the original reasons for doing things with multiple subimages no
no longer applied, so I changed things.