Adjust code that places blocks to work with sprite rotation
The check was occurring at the origin position of a given location
in the grid; however, this position would still be occupied by the
sprite, even after a rotation.

Instead, make the check in the middle of the grid positions, so that
slight differences based on rotations won't cause false positive detections
or false negative misses.

Also remove instances of the block grid from the code. It may not be
used after all